+++ ApiProyectoJava/src/main/java/edu/apiProyectoJava/servicios/IncidenciaRepositorio.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import edu.apiProyectoJava.bean.Incidencia;
 import edu.apiProyectoJava.bean.Usuario;
 
-public interface IncidenciaRepositorio extends JpaRepository<Incidencia, Long>{
+public interface IncidenciaRepositorio extends JpaRepository<Incidencia, Integer>{
 
 }
