+++ ApiProyectoJava/src/main/java/edu/apiProyectoJava/servicios/SolicitudRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import edu.apiProyectoJava.bean.Solicitud;
 
-public interface SolicitudRepositorio extends JpaRepository<Solicitud, Long>{
+public interface SolicitudRepositorio extends JpaRepository<Solicitud, Integer>{
 
 }
